+++ 1.20/Fabric/src/main/java/fuzs/puzzleslib/mixin/MobFabricMixin.java
@@ -5,6 +5,7 @@ import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.puzzleslib.api.event.v1.data.DefaultedValue;
 import fuzs.puzzleslib.impl.event.SpawnDataMob;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.DifficultyInstance;
 import net.minecraft.world.entity.*;
 import net.minecraft.world.level.Level;
@@ -16,6 +17,7 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(Mob.class)
@@ -47,5 +49,18 @@ abstract class MobFabricMixin extends LivingEntity implements SpawnDataMob {
         DefaultedValue<LivingEntity> target = DefaultedValue.fromValue(entity);
         EventResult result = FabricLivingEvents.LIVING_CHANGE_TARGET.invoker().onLivingChangeTarget(this, target);
         return result.isInterrupt() ? this.target : target.getAsOptional().orElse(entity);
+    }
+
+    @Inject(method = "checkDespawn", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/Level;getNearestPlayer(Lnet/minecraft/world/entity/Entity;D)Lnet/minecraft/world/entity/player/Player;", shift = At.Shift.BEFORE), cancellable = true)
+    public void checkDespawn(CallbackInfo callback) {
+        EventResult result = FabricLivingEvents.CHECK_MOB_DESPAWN.invoker().onCheckMobDespawn(Mob.class.cast(this), (ServerLevel) this.level());
+        if (result.isInterrupt()) {
+            if (result.getAsBoolean()) {
+                this.discard();
+            } else {
+                this.noActionTime = 0;
+            }
+            callback.cancel();
+        }
     }
 }